Create README.md
Images are all around us, but until recently it has been difficult and expensive to try and analyze them. Computer vision tools have made this easier, and a variety of open source AI models are available to let you provide useful reasoning about the contents of images.You'll use the state-of-the-art vision transformer model to classify images and detect objects within images. You'll also learn how to provide labeled examples to improve the quality of the AI's responses.